tests: consolidate kcrypt e2e specs into monorepo tests/
Ports the kcrypt-challenger e2e coverage that is not already in the
monorepo suite, drops the specs that duplicated existing coverage,
and deletes the kcrypt-side tree they came from along with the
pre-monorepo workflow that used to drive them.

Kept (ported to tests/encryption_advanced_test.go):

  * remote-complete-workflow -- the monolithic TOFU + quarantine +
    PCR/AK management + secret-reuse + multi-partition spec. Ported
    verbatim as a single It; splitting it would double the wall clock
    for no coverage gain, since each phase reuses the same VM boot
    and enrollment state.
  * remote-ek-only, remote-selective-pcr, remote-ek-reenroll,
    remote-mixed-modes -- the selective-enrollment specs. Each is a
    distinct scenario and was already its own Describe upstream; the
    labels are prefixed with `encryption-` to line up with the
    monorepo cell-selection convention (`.github/encryption-tests.sh`
    invokes `ginkgo --label-filter "$LABEL"` and the reusable qemu
    workflow branches on `startsWith(inputs.test, 'encryption-')`).

Dropped (already covered by tests/encryption_test.go):

  * local-encryption, remote-auto, remote-static,
    remote-https-pinned, remote-https-bad-cert -- one-to-one with
    the monorepo suite's encryption-local, encryption-remote-auto,
    encryption-remote-static, encryption-remote-https-pinned, and
    encryption-remote-https-bad-cert cells.
  * discoverable-kms -- was `XWhen`-disabled upstream with a TODO
    about qemu bridge networking; not a regression to drop.

Helpers copied over next to the specs: installKairosWithConfigAdvanced,
rebootAndConnect, verifyEncryptedPartition, checkPassphraseRetrieval,
expectPassphraseRetrieval, expectPassphraseRetrievalWithError,
createSealedVolumeWithAttestation, updateSealedVolumeAttestation,
quarantineTPM, unquarantineTPM, deleteSealedVolume, secretExists,
cleanupTestResources. getSealedVolumeName is inlined -- the input is
always `tofu-<8 lowercase hex chars>`, 13 chars total, so the upstream
SafeKubeName truncation path is never taken and the helper reduces to
sprintf.

New CI cells added to qemu-tests-standard-encryption in pr.yaml,
master.yaml, and release.yaml:
  - encryption-remote-complete-workflow
  - encryption-remote-ek-only
  - encryption-remote-selective-pcr
  - encryption-remote-ek-reenroll
  - encryption-remote-mixed-modes

Deletions in kcrypt/:
  * tests/ (whole directory: 4 spec files + 4 asset YAMLs).
  * scripts/ (only e2e-tests.sh, the wrapper .github/encryption-tests.sh
    replaces at repo root).
  * Dockerfile.kairos-image (built the pre-monorepo test ISO by cloning
    the archived immucore + kairos-agent repos; only reachable through
    the dead kcrypt/.github/workflows/e2e-tests.yml).
  * .github/ (dead pre-monorepo workflows and issue template; GitHub
    only reads .github/ at the repo root).
  * Makefile: dropped the "Kairos Image Build" section (kairos-image /
    kairos-iso / kairos-iso-uki / genkeys plus their build-arg
    variables) and the "Cloud Init Datasource ISO" section
    (datasource-iso). The kubebuilder scaffolding targets are left in
    place; they are a distinct cleanup.

Refs kairos-io/kairos#4367.
