--- kcrypt/.github/ISSUE_TEMPLATE/file-issues-on-main-kairos-repo.md
@@ -1,12 +1,0 @@
----
-name: File issues on main Kairos repo
-about: Tell users to file their issues on the main Kairos repo
-title: ''
-labels: ''
-assignees: ''
-
----
-
-:warning: All Kairos issues are tracked in our main repo, please file your issue there, thanks! :warning:
-
-https://github.com/kairos-io/kairos/issues
