--- kcrypt/.github/workflows/lint.yml
@@ -1,28 +1,0 @@
-name: Lint
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    paths:
-      - '**'
-
-
-concurrency:
-  group: ci-lint-${{ github.head_ref || github.ref }}-${{ github.repository }}
-  cancel-in-progress: true
-
-env:
-  FORCE_COLOR: 1
-jobs:
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v7
-        with:
-          fetch-depth: 0
-      - name: Run yamllint
-        run: |
-          pip install yamllint
-          yamllint -c .yamllint .github/workflows/
