--- kcrypt/.github/workflows/osv-scanner-pr.yaml
@@ -1,69 +1,0 @@
-name: OSV-Scanner PR Scan
-
-on:
-  pull_request:
-  push:
-    branches:
-      - main
-  merge_group:
-    branches: [main]
-
-permissions:
-  security-events: write
-  contents: read
-  actions: read
-
-jobs:
-  scan-pr:
-    runs-on: ubuntu-latest
-    env:
-      OSV_SCANNER_VERSION: v2.4.0
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1  # v7.0.1
-        with:
-          persist-credentials: false
-
-      - name: Install Go
-        uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e  # v7.0.0
-        with:
-          go-version-file: go.mod
-
-      - name: Install osv-scanner
-        run: |
-          go install "github.com/google/osv-scanner/v2/cmd/osv-scanner@${OSV_SCANNER_VERSION}"
-          osv-scanner --version
-
-      - name: Run osv-scanner
-        id: scan
-        continue-on-error: true
-        run: |
-          osv-scanner scan source \
-            --recursive \
-            --format=sarif \
-            --output=results.sarif \
-            ./
-
-      - name: Upload SARIF artifact
-        if: ${{ !cancelled() && hashFiles('results.sarif') != '' }}
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a  # v7.0.1
-        with:
-          name: OSV Scanner SARIF file
-          path: results.sarif
-          retention-days: 5
-
-      - name: Upload SARIF to code-scanning
-        if: ${{ !cancelled() && hashFiles('results.sarif') != '' }}
-        uses: github/codeql-action/upload-sarif@5595ccaf912efad79be6eef63a5619ff05969be3  # v4.37.6
-        with:
-          sarif_file: results.sarif
-
-      - name: Report scan outcome
-        if: steps.scan.outcome == 'failure'
-        run: |
-          if [ -f results.sarif ]; then
-            echo "osv-scanner reported vulnerabilities; see the Security tab or the results.sarif artifact."
-          else
-            echo "osv-scanner failed before writing results.sarif; see the 'Run osv-scanner' step logs."
-          fi
-          exit 1
