--- kcrypt/.github/workflows/unit-tests.yml
@@ -1,48 +1,0 @@
-name: Unit tests
-on:
-  push:
-    branches:
-      - master
-  pull_request:
-env:
-  FORCE_COLOR: 1
-concurrency:
-  group: ci-unit-${{ github.head_ref || github.ref }}-${{ github.repository }}
-  cancel-in-progress: true
-jobs:
-  unit-tests:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v7
-        with:
-          fetch-depth: 0
-      - name: Install Go
-        uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7.0.0
-        with:
-          go-version-file: go.mod
-      - name: Cache Go modules
-        uses: actions/cache@v6
-        with:
-          path: ~/go/pkg/mod
-          key: ${{ runner.os }}-go-${{ hashFiles('**/go.sum') }}
-          restore-keys: |
-            ${{ runner.os }}-go-
-      - name: Install dependencies
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y libssl-dev
-      - name: Generate manifests and code
-        run: |
-          make manifests generate
-      - name: Run tests
-        env:
-          CGO_ENABLED: 1
-        run: |
-          go run github.com/onsi/ginkgo/v2/ginkgo run --covermode=atomic --coverprofile=coverage.out -p -r pkg/challenger cmd/discovery/client
-      - name: Codecov
-        uses: codecov/codecov-action@v7
-        env:
-          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
-        with:
-          file: ./coverage.out
