+++ kcrypt/Makefile
@@ -260,98 +260,3 @@ kubesplit: manifests kustomize
 	mkdir helm-chart
 	$(KUSTOMIZE) build config/default | kubesplit -helm helm-chart
 
-##@ Kairos Image Build
-
-# Kairos image build variables
-KAIROS_IMAGE_TAG ?= kcrypt-challenger:latest
-AURORABOOT_IMAGE ?= quay.io/kairos/auroraboot
-ISO_NAME ?= challenger
-BUILD_DIR = build
-KEYS_DIR ?= $(HOME)/tmp/keys
-
-# Optional build args for custom binaries (empty = use package versions)
-# Set REF to build custom binaries from kairos-io repos (ref can be branch or commit)
-IMMUCORE_REF ?=
-KAIROS_AGENT_REF ?=
-KAIROS_INIT_VERSION ?= v0.5.20
-NO_CACHE ?=
-
-.PHONY: kairos-image
-kairos-image: ## Build Kairos image with challenger
-	@echo "Building Kairos image..."
-	@echo "Optional: Set IMMUCORE_REF or KAIROS_AGENT_REF to build custom binaries"
-	docker build --progress=plain $(if $(NO_CACHE),--no-cache) -f Dockerfile.kairos-image -t $(KAIROS_IMAGE_TAG) \
-		$(if $(IMMUCORE_REF),--build-arg IMMUCORE_REF=$(IMMUCORE_REF)) \
-		$(if $(KAIROS_AGENT_REF),--build-arg KAIROS_AGENT_REF=$(KAIROS_AGENT_REF)) \
-		--build-arg KAIROS_INIT_VERSION=$(KAIROS_INIT_VERSION) \
-		.
-
-.PHONY: genkeys
-genkeys: ## Generate secure boot keys for UKI signing
-	@echo "Generating keys in $(KEYS_DIR)..."
-	@mkdir -p $(KEYS_DIR)
-	@docker run --rm \
-		-v "$(KEYS_DIR):/work/keys" \
-		$(AURORABOOT_IMAGE) \
-		genkey --skip-microsoft-certs-I-KNOW-WHAT-IM-DOING --expiration-in-days 365 -o /work/keys "$(KEYS_ORG)"
-	@echo "Keys generated successfully in $(KEYS_DIR)"
-
-.PHONY: kairos-iso
-kairos-iso: kairos-image ## Build normal Kairos ISO using auroraboot
-	@echo "Building ISO using auroraboot..."
-	@mkdir -p $(BUILD_DIR)
-	@docker run --rm \
-		-v /var/run/docker.sock:/var/run/docker.sock \
-		-v "$(CURDIR)/$(BUILD_DIR):/tmp/auroraboot" \
-		$(AURORABOOT_IMAGE) \
-		build-iso $(KAIROS_IMAGE_TAG)
-	@echo "ISO built successfully in $(BUILD_DIR)/"
-	@ls -lh $(BUILD_DIR)/*.iso 2>/dev/null || echo "Note: ISO files may have different naming"
-
-.PHONY: kairos-iso-uki
-kairos-iso-uki: kairos-image ## Build UKI (Unified Kernel Image) ISO with secure boot signing
-	@echo "Building UKI ISO using auroraboot..."
-	@mkdir -p $(BUILD_DIR)
-	@if [ ! -d "$(KEYS_DIR)" ] || [ ! -f "$(KEYS_DIR)/db.key" ]; then \
-		echo "Error: Keys directory not found or missing keys at $(KEYS_DIR)"; \
-		echo "Keys are required for UKI ISO build. Run 'make genkeys' to generate them."; \
-		exit 1; \
-	fi
-	@echo "Building UKI with secure boot signing..."; \
-	docker run --rm \
-		-v /var/run/docker.sock:/var/run/docker.sock \
-		-v "$(CURDIR)/$(BUILD_DIR):/result" \
-		-v "$(KEYS_DIR):/keys" \
-		$(AURORABOOT_IMAGE) \
-		build-uki -t iso -d /result/ \
-		--public-keys /keys \
-		--tpm-pcr-private-key /keys/tpm2-pcr-private.pem \
-		--sb-key /keys/db.key \
-		--sb-cert /keys/db.pem \
-		--extend-cmdline "console=tty0 console=ttyS0,115200n8 rd.immucore.debug" \
-		$(KAIROS_IMAGE_TAG)
-	@echo "UKI ISO built successfully in $(BUILD_DIR)/"
-	@ls -lh $(BUILD_DIR)/*.iso 2>/dev/null || echo "Note: ISO files may have different naming"
-
-		#--extend-cmdline "console=tty0 console=ttyS0,115200n8 rd.immucore.debug kcrypt.challenger.challenger_server=http://192.168.122.1.challenger.sslip.io" \
-
-##@ Cloud Init Datasource ISO
-
-# Build cloud-init datasource ISO (for cloud-init NoCloud datasource)
-# Usage: make datasource-iso CLOUD_CONFIG=tests/assets/qrcode.yaml
-.PHONY: datasource-iso
-datasource-iso: ## Build cloud-init datasource ISO for NoCloud
-	@if [ -z "$(CLOUD_CONFIG)" ]; then \
-		echo "Error: CLOUD_CONFIG is required. Usage: make datasource-iso CLOUD_CONFIG=path/to/config.yaml"; \
-		exit 1; \
-	fi
-	@mkdir -p $(BUILD_DIR)
-	@docker run --rm \
-		-v "$(CURDIR)/$(BUILD_DIR):/build" \
-		-v "$(CURDIR)/$(CLOUD_CONFIG):/tmp/user-data:ro" \
-		quay.io/kairos/osbuilder-tools \
-		sh -c "zypper in -y mkisofs && \
-		       touch /build/meta-data && \
-		       cp /tmp/user-data /build/user-data && \
-		       mkisofs -output /build/ci.iso -volid cidata -joliet -rock /build/user-data /build/meta-data"
-	@echo "Datasource ISO built: $(BUILD_DIR)/ci.iso"
