--- kcrypt/tests/assets/cluster-issuer.yaml
@@ -1,8 +1,0 @@
----
-# Self-signed issuer
-apiVersion: cert-manager.io/v1
-kind: ClusterIssuer
-metadata:
-  name: selfsigned
-spec:
-  selfSigned: {}
