--- kcrypt/tests/assets/kustomization.yaml
@@ -1,13 +1,0 @@
-# Adds namespace to all resources.
-namespace: default
-
-bases:
-- ../../config/default
-
-resources:
-- challenger-server-ingress.yaml
-- cluster-issuer.yaml
-
-patchesStrategicMerge:
-# Fix labels and selectors to make challenger server accessible
-- challenger-patch.yaml
